Modify Post class and folder reorganisation
